Bugix | Refactoring code and structure, made dashboard app to storage company dashnoard functions
