+++ base/accounts/urls.py
@@ -41,11 +41,6 @@ urlpatterns = [
         name='success_password_change'
     ),
     path(
-        'dashboard/',
-        views.CompanyDashboard.as_view(),
-        name='dashboard'
-    ),
-    path(
         'profile/update/<int:pk>/',
         views.ProfileUpdateView.as_view(),
         name='update_profile'
@@ -56,10 +51,6 @@ urlpatterns = [
         name='user_profile'
     ),
 
-    path(
-        'company/add-review/<int:company_id>/',
-        views.AddCompanyReviewView.as_view(),
-        name='add_review'
-    )
+
 
 ]
