+++ base/base/templates/navigation_bar.html
@@ -38,7 +38,7 @@
           <ul class="navbar-nav me-auto mb-2 mb-lg-0">
             <li class="nav-item" style="margin-right: 10px">
               {% if user.role == 'company' %}
-                <a style="color: black; font-weight: 600" class="nav-link" href="{% url 'accounts:dashboard' %}"> Dashboard </a>
+                <a style="color: black; font-weight: 600" class="nav-link" href="{% url 'dashboard:dashboard' %}"> Dashboard </a>
                 {% elif user.role == 'user' %}
                 <a style="color: black; font-weight: 600" class="nav-link" href="{% url 'accounts:user_profile' %}"> Profile </a>
               {% endif %}
