+++ base/dashboard/templates/applications_list.html
