+++ base/dashboard/templates/company_dashboard.html
@@ -7,7 +7,7 @@
 
 
 <div class="container">
-    <a class="btn btn-primary" role="button" href="{% url 'accounts:create' %}">Add new offer</a>
+    <a class="btn btn-primary" role="button" href="{% url 'dashboard:create-offer' %}">Add new offer</a>
     <a class="btn btn-primary" role="button" href="{% url 'accounts:update_profile' user.id %}">Edit company information</a>
 </div>
 
@@ -22,13 +22,13 @@
                 </div>
                   <div class="d-flex justify-content-between">
                     <p class="card-text"><small class="text-body-secondary">{{object.date_created}}</small></p>
-                    <a class="card-text" href="{% url 'accounts:applications' object.id %}">Candidates</a>
-                    <a class="card-text" href="{% url 'accounts:edit' object.id %}">Edit</a>
-                    <a class="card-text" href="{% url 'accounts:generate_csv' object.id %}">
+                    <a class="card-text" href="{% url 'dashboard:applications' object.id %}">Candidates</a>
+                    <a class="card-text" href="{% url 'dashboard:edit-offer' object.id %}">Edit</a>
+                    <a class="card-text" href="{% url 'dashboard:generate-csv' object.id %}">
                         CSV
                     </a>
                     
-                      <form method="POST" action="{% url 'accounts:delete' object.id %}">
+                      <form method="POST" action="{% url 'dashboard:delete-offer' object.id %}">
                           {% csrf_token %}
                           <button style="margin-left: 10px" type="submit" class="btn btn-danger">DELETE</button>
                       </form>
