+++ base/offers/urls.py
@@ -13,22 +13,22 @@ urlpatterns = [
     path(
         'companies/',
         views.CompaniesListView.as_view(),
-        name='companies'
+        name='companies-list'
     ),
     path(
         'offer/<int:pk>/',
         views.OfferDetailView.as_view(),
-        name='offer'
+        name='offer-detail'
     ),
     path(
         'company/<int:pk>/',
         views.CompanyDetailView.as_view(),
-        name='company'
+        name='company-detail'
     ),
     path(
         'apply/<int:offer_id>/',
         views.ApplyForOfferView.as_view(),
-        name='apply'
+        name='apply-offer'
     ),
     path(
         'apply/success',
@@ -36,39 +36,9 @@ urlpatterns = [
         name='apply-success'
     ),
     path(
-        'offer/edit/<int:pk>/',
-        views.OfferUpdateView.as_view(),
-        name='edit'
-    ),
-    path(
-        'offer/create/',
-        views.OfferCreateView.as_view(),
-        name='create'
-    ),
-    path(
-        'offer/delete/<int:pk>/',
-        views.OfferDeleteView.as_view(),
-        name='delete'
-    ),
-    path(
-        'applications/<int:offer_id>/',
-        views.ApplicationsListView.as_view(),
-        name='applications'
-    ),
-    path(
-        'application/sendfeedback/<int:application_id>/',
-        views.ReturnApplicationFeedbackView.as_view(),
-        name='send_feedback'
-    ),
-    path(
-        'application/delete/<int:pk>/',
-        views.ApplicationDeleteView.as_view(),
-        name='delete_application'
-    ),
-    path(
-        'application/generate-csv/<int:pk>/',
-        views.generate_application_csv,
-        name='generate_csv'
-    ),
+        'company/add-review/<int:company_id>/',
+        views.AddCompanyReviewView.as_view(),
+        name='add-review'
+    )
 
 ]
