+++ base/study/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
+
 from .models import Category, Resources
