+++ base/study/forms.py
@@ -1,5 +1,6 @@
 from django import forms
 from django.forms import ModelMultipleChoiceField
+
 from .models import Category
